product-service/requirements.txt
Django==4.2.27
djangorestframework==3.15.2
psycopg2-binary==2.9.10
requests==2.32.3
gunicorn==23.0.0
django-cors-headers==4.6.0
python-dotenv==1.0.1
